+++ pfcpType/MeasurementInformation.go
@@ -15,7 +15,7 @@ func (m *MeasurementInformation) MarshalBinary() (data []byte, err error) {
 	tmpUint8 := btou(m.Radi)<<2 |
 		btou(m.Inam)<<1 |
 		btou(m.Mbqe)
-	data = append([]byte(""), byte(tmpUint8))
+	data = append([]byte(""), tmpUint8)
 
 	return data, nil
 }
@@ -28,9 +28,9 @@ func (m *MeasurementInformation) UnmarshalBinary(data []byte) error {
 	if length < idx+1 {
 		return fmt.Errorf("Inadequate TLV length: %d", length)
 	}
-	m.Radi = utob(uint8(data[idx]) & BitMask3)
-	m.Inam = utob(uint8(data[idx]) & BitMask2)
-	m.Mbqe = utob(uint8(data[idx]) & BitMask1)
+	m.Radi = utob(data[idx] & BitMask3)
+	m.Inam = utob(data[idx] & BitMask2)
+	m.Mbqe = utob(data[idx] & BitMask1)
 	idx = idx + 1
 
 	if length != idx {
