+++ pfcpType/NodeReportType.go
@@ -11,7 +11,7 @@ type NodeReportType struct {
 func (n *NodeReportType) MarshalBinary() (data []byte, err error) {
 	// Octet 5
 	tmpUint8 := btou(n.Upfr)
-	data = append([]byte(""), byte(tmpUint8))
+	data = append([]byte(""), tmpUint8)
 
 	return data, nil
 }
@@ -24,7 +24,7 @@ func (n *NodeReportType) UnmarshalBinary(data []byte) error {
 	if length < idx+1 {
 		return fmt.Errorf("Inadequate TLV length: %d", length)
 	}
-	n.Upfr = utob(uint8(data[idx]) & BitMask1)
+	n.Upfr = utob(data[idx] & BitMask1)
 	idx = idx + 1
 
 	if length != idx {
