+++ pfcpType/PFCPSMReqFlags.go
@@ -15,7 +15,7 @@ func (p *PFCPSMReqFlags) MarshalBinary() (data []byte, err error) {
 	tmpUint8 := btou(p.Qaurr)<<2 |
 		btou(p.Sndem)<<1 |
 		btou(p.Drobu)
-	data = append([]byte(""), byte(tmpUint8))
+	data = append([]byte(""), tmpUint8)
 
 	return data, nil
 }
@@ -28,9 +28,9 @@ func (p *PFCPSMReqFlags) UnmarshalBinary(data []byte) error {
 	if length < idx+1 {
 		return fmt.Errorf("Inadequate TLV length: %d", length)
 	}
-	p.Qaurr = utob(uint8(data[idx]) & BitMask3)
-	p.Sndem = utob(uint8(data[idx]) & BitMask2)
-	p.Drobu = utob(uint8(data[idx]) & BitMask1)
+	p.Qaurr = utob(data[idx] & BitMask3)
+	p.Sndem = utob(data[idx] & BitMask2)
+	p.Drobu = utob(data[idx] & BitMask1)
 	idx = idx + 1
 
 	if length != idx {
