+++ pfcpType/PFCPSRRspFlags.go
@@ -11,7 +11,7 @@ type PFCPSRRspFlags struct {
 func (p *PFCPSRRspFlags) MarshalBinary() (data []byte, err error) {
 	// Octet 5
 	tmpUint8 := btou(p.Drobu)
-	data = append([]byte(""), byte(tmpUint8))
+	data = append([]byte(""), tmpUint8)
 
 	return data, nil
 }
@@ -24,7 +24,7 @@ func (p *PFCPSRRspFlags) UnmarshalBinary(data []byte) error {
 	if length < idx+1 {
 		return fmt.Errorf("Inadequate TLV length: %d", length)
 	}
-	p.Drobu = utob(uint8(data[idx]) & BitMask1)
+	p.Drobu = utob(data[idx] & BitMask1)
 	idx = idx + 1
 
 	if length != idx {
