+++ pfcpType/ReportingTriggers.go
@@ -31,7 +31,7 @@ func (r *ReportingTriggers) MarshalBinary() (data []byte, err error) {
 		btou(r.Timth)<<2 |
 		btou(r.Volth)<<1 |
 		btou(r.Perio)
-	data = append([]byte(""), byte(tmpUint8))
+	data = append([]byte(""), tmpUint8)
 
 	// Octet 6
 	tmpUint8 = btou(r.Evequ)<<5 |
@@ -40,7 +40,7 @@ func (r *ReportingTriggers) MarshalBinary() (data []byte, err error) {
 		btou(r.Envcl)<<2 |
 		btou(r.Timqu)<<1 |
 		btou(r.Volqu)
-	data = append(data, byte(tmpUint8))
+	data = append(data, tmpUint8)
 
 	return data, nil
 }
@@ -53,26 +53,26 @@ func (r *ReportingTriggers) UnmarshalBinary(data []byte) error {
 	if length < idx+1 {
 		return fmt.Errorf("Inadequate TLV length: %d", length)
 	}
-	r.Liusa = utob(uint8(data[idx]) & BitMask8)
-	r.Droth = utob(uint8(data[idx]) & BitMask7)
-	r.Stopt = utob(uint8(data[idx]) & BitMask6)
-	r.Start = utob(uint8(data[idx]) & BitMask5)
-	r.Quhti = utob(uint8(data[idx]) & BitMask4)
-	r.Timth = utob(uint8(data[idx]) & BitMask3)
-	r.Volth = utob(uint8(data[idx]) & BitMask2)
-	r.Perio = utob(uint8(data[idx]) & BitMask1)
+	r.Liusa = utob(data[idx] & BitMask8)
+	r.Droth = utob(data[idx] & BitMask7)
+	r.Stopt = utob(data[idx] & BitMask6)
+	r.Start = utob(data[idx] & BitMask5)
+	r.Quhti = utob(data[idx] & BitMask4)
+	r.Timth = utob(data[idx] & BitMask3)
+	r.Volth = utob(data[idx] & BitMask2)
+	r.Perio = utob(data[idx] & BitMask1)
 	idx = idx + 1
 
 	// Octet 6
 	if length < idx+1 {
 		return fmt.Errorf("Inadequate TLV length: %d", length)
 	}
-	r.Evequ = utob(uint8(data[idx]) & BitMask6)
-	r.Eveth = utob(uint8(data[idx]) & BitMask5)
-	r.Macar = utob(uint8(data[idx]) & BitMask4)
-	r.Envcl = utob(uint8(data[idx]) & BitMask3)
-	r.Timqu = utob(uint8(data[idx]) & BitMask2)
-	r.Volqu = utob(uint8(data[idx]) & BitMask1)
+	r.Evequ = utob(data[idx] & BitMask6)
+	r.Eveth = utob(data[idx] & BitMask5)
+	r.Macar = utob(data[idx] & BitMask4)
+	r.Envcl = utob(data[idx] & BitMask3)
+	r.Timqu = utob(data[idx] & BitMask2)
+	r.Volqu = utob(data[idx] & BitMask1)
 	idx = idx + 1
 
 	if length != idx {
